Apply code review change requests.
