+++ inference-engine/tests/functional/plugin/gpu/shared_tests_instances/behavior/set_blob_of_kind.cpp
@@ -8,7 +8,11 @@
 using namespace BehaviorTestsDefinitions;
 using namespace InferenceEngine;
 
-const std::vector<BlobKind> blobKinds{BlobKind::Simple, BlobKind::Compound, BlobKind::BatchOfSimple};
+const std::vector<FuncTestUtils::BlobKind> blobKinds = {
+    FuncTestUtils::BlobKind::Simple,
+    FuncTestUtils::BlobKind::Compound,
+    FuncTestUtils::BlobKind::BatchOfSimple
+};
 
 const SetBlobOfKindConfig gpuConfig{}; //nothing special
